requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and LLM
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0

# Optimization
ortools==9.8.0

# Data handling
pandas==2.1.0
numpy==1.26.0
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Optional
streamlit==1.30.0
plotly==5.18.0

# Development
pytest==7.4.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
black==24.1.0
flake8==7.0.0
mypy==1.8.0
